Split git_export into 3 modules
